test making and reading custom Eyelink classes
The notion of `experiment` is custom to how I run my experiments (the first
message sent always sends along the filename for the experiment). This commit
adds a test_customeyelink.py file which contains a subclass of the
`pyarbus.data.Eyelink` class that knows how to extract the experiment name
from the first message, and shows how to read it.